requirements.txt
fastapi[all]
gunicorn
uvicorn
jinja2
pytest
requests
aiofiles
python-multipart
pillow
